outlying class for peak and trough
